+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * Add `breaker:intrinsic` search that highlights items that have an intrinsic breaker ability (i.e. not granted by the seasonal artifact).
+* Breaker type granted by the seasonal artifact now has a green box around it, reminiscent of the artifact mod that grants it.
 
 ## 8.37.0 <span class="changelog-date">(2024-09-15)</span>